fixed Unix domain socket bind address already in use problem

+++ engine_race/format/format.cc
@@ -16,7 +16,7 @@ void qfmtColorizerF(FILE *f, const char *prefix, const char *colored, const char
 
 std::string KVArrayDump(const char* arr, size_t len){
     std::string s = "[";
-    for(int i = 0; i < len; i+=sizeof(uint16_t)){
+    for(size_t i = 0; i < len; i+=sizeof(uint16_t)){
         char tmp[10] = {0};
         sprintf(tmp, "%hu", *reinterpret_cast<const uint16_t*>(arr + i));
         if(i + 2 < len){
